+++ src/pages/notice/abnormalTrade.tsx
@@ -2,12 +2,16 @@ import { Accordion, Box, Flex, Image } from '@chakra-ui/react';
 import { GetStaticProps } from 'next';
 import { useTranslation } from 'next-i18next';
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
+import dynamic from 'next/dynamic';
 import { useRouter } from 'next/router';
 import React from 'react';
-import InfoAccordionItem from '../../components/Common/InfoAccordionItem';
 import InfoTitle from '../../components/Common/InfoTitle';
 import Wrapper from '../../components/Wrapper';
 import { Locales } from '../../i18n/locales';
+const InfoAccordionItem = dynamic(
+  () => import('../../components/Common/InfoAccordionItem'),
+  { ssr: false }
+);
 
 const abnormalTrade: React.FC<{}> = () => {
   const { t } = useTranslation(['notice']);