+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.1.0](https://github.com/alice1qaz/WCG/compare/v1.0.1...v1.1.0) (2021-07-24)
+
+
+### Features
+
+* **all:** animationsOnScroll ([a500b0b](https://github.com/alice1qaz/WCG/commit/a500b0bff92ba786c53d51d21dbf51ad6a2fa950))
+* **server:** add server log function ([e9b163a](https://github.com/alice1qaz/WCG/commit/e9b163a189e240e2665c60cf2939e29dd9884e9d))
+
+
+### Bug Fixes
+
+* **infoaccordionitem:** use dynamic import for InfoAccordionItem ([6b52ec8](https://github.com/alice1qaz/WCG/commit/6b52ec8ebb2d102c83cf130b027c83597bc385a3))
+* **products:** hide cryptocurrencies pages ([2e96291](https://github.com/alice1qaz/WCG/commit/2e96291427a24501f4598610773fc951a9fd924b))
+
 ### [1.0.3](https://github.com/alice1qaz/WCG/compare/v1.0.2...v1.0.3) (2021-07-23)